Fixed int -> uint16 conversion warning

+++ examples/src/example_image_publisher.cpp
@@ -54,7 +54,8 @@ int main() {
     // Set the image dimensions (e.g. 512x512x1).
     img.setImageDimensions(image.cols, image.rows, 1);
     // Set the image data, giving also the total size in bytes and the number of channels
-    img.setImageData(image.data, static_cast<int>(image.total() * image.elemSize()), image.channels());
+    img.setImageData(image.data, static_cast<int>(image.total() * image.elemSize()),
+                     static_cast<int>(image.channels()));
     // Publish the image.
     pub.publish(img);
